Split cache specs into their own file
